Implement complexityChecks, enforce in strong mode
This implement basic complexity checks to help prevent usage
of schemas that are likely to be exploitable to DoS attacks
on untrusted input.

This is not _very_ smart, i.e. it just checks that maxLength
is present where needed but would accept a giant maxLength,
and also it requires all regexps with two stars to be limited,
depite that a significant part of them are actually fine.

safe-regex is not used due to known false negatives/positives,
instead a very simplistic conservative check is implemented.

Tests included.