extract sub-schemas with fully resolved ids from all schemas
This allows to pass a sub-schema with a sub-schema having a fully
resolved $id, which will be added as a new top-level schema.

This is not ideal as works only with full ids, but this makes the
solution less complex.

This is supported only in cases where schemas are passed as an array
for their ids to be extracted from the schemas themselves.

Also, this changes the behavior of #93 to support only fully resolved
ids.

Tests will be introduced separately.