Make errors from $ref validation transparent
This approach is completely different from what ajv does.

This sees "through" $refs, building full error references from the top,
and descending into refs like '/something/$ref/path/continues' instead
of replacing the whole path with the path where the reference resolves
to.

This way, it is actually possible to tell where the error originated
from, unlike with ajv which makes it hard to understand which schema
caused the error in some situations.