Split validator() from compile() (#110)
* Split validator() from compile()

This allows us to add top-level preparation separately, instead of
repeating it on every subschema ref compilation or having to use hacks
to exclude preparation code from re-running.

I.e. buildSchemas is now called only once, and json check is now
explicitly moved to a wrapping function.

That will likely be used to add some other common preparation steps that
we will need to run only at top-level.

* Don't need ...more in error() function now

This comes with a ~10x speedup on verboseErrors.