parser: embed first error in concise form if errors are enabled
If full errors are needed (and enabled), they can be accessed via the
the 'errors' property, as documented.

No reason to attempt to format all of them into the error message, that
was both not pretty and broken, as 'message' was removed and those
didn't include 'dataPath'.

Also don't set undefined 'errors' property if errors are not enabled.