Improve static evaluated tracing
Dynamic is no longer an irreversable flag, now definitely evaluating
property which was maybe evaluated can reverse the dynamic status.

That is done by keeping track of evaluated and potentially evaluated
and accounting to that while doing merges and intersections.

isDynamic() check now checks that all potentially evaluated are also
definitely evaluated.

This is also now separately for items and properties, dynamic properties
alone won't trigger a re-run with dynamic tracing if unevaluatedItems
is used and not unevaluatedProperties, for example.