implement requireSchema and enable in strong mode
Thus, other implementations would be less likely to break when ignoring
keywords that are present in the schema but they don't support.

E.g. $ref alongside other keywords can be silently ignored in some impls
if $schema is not specified, but that could be fixed in some cases by
specifying a $schema. In that case, other implementations are more
likely to either follow the correct action or throw a visible error
about schema version not being supported.