feat: implement { removeAdditional: 'keyword' } option
It enables a non-standard 'removeAdditional' keyword for fine-grained
opt-in, and will remove additional items/properites only in those paths,
where it's explicitly enabled by the { removeAdditional: true } keyword.

The only supported value for the 'removeAdditional' keyword is true.

If { removeAdditional: 'keyword' } is not enabled, using
'removeAdditional' keyword is an error.