chore: simplify tracing of known type and required in nested rules
E.g. in allOf/oneOf, we just rely on the checks in parent rule,
but only for type/required checks.

In the following schema, the nested allOf will just check
for `{ required: ['yyy'] }`:

```
{
  type: 'object',
  required: ['xxx','zzz'],
  allOf: [{
    type: 'object',
    required: ['xxx', 'yyy']
  }]
}
```

Properties/items can't be inherited as they affect evaluation checks,
which should be isolated from the parent.